refactor: 외부 API에 대한 의존성 줄임
